requirements.txt
opencv-python>=4.8.0
Pillow>=10.1.0
pyyaml>=6.0
colorama>=0.4.6
tqdm>=4.66.0
psutil>=5.9.0 